refactor(ethexe/processor): split run module into submodules (#5120)
- Split ethexe/processor/src/handling/run.rs (1006 lines) into separate module
- Create run/ directory with logical components:
  - mod.rs: Main execution loop, RunContext trait, CommonRunContext
  - chunks_splitting.rs: Chunk partitioning algorithm and data structures
  - chunk_execution_spawn.rs: Thread pool spawning and parallel execution
  - chunk_execution_processing.rs: Journal collection and processing
- Update overlaid.rs imports to use re-exported types
- All 18 tests pass, no functional changes
