+++ ethexe/processor/src/handling/run/mod.rs
@@ -100,17 +100,18 @@
 //!
 //! ## Future Improvements
 //!
-//! Currently, the chunk partitioning algorithm is simple and does not consider a program’s execution time.
+//! Currently, the chunk partitioning algorithm is simple and does not consider a program's execution time.
 //! In the future, we could introduce a weight multiplier to the queue size to improve partitioning efficiency.
 //! This weight multiplier could be calculated based on program execution time statistics.
 
-// TODO: #5120 split to several files and move to separate module
+mod chunk_execution_processing;
+mod chunk_execution_spawn;
+pub(super) mod chunks_splitting;
 
-use crate::{
-    ProcessorError, Result, handling::run::chunks_splitting::ExecutionChunks, host::InstanceCreator,
-};
+pub(crate) use chunks_splitting::{ActorStateHashWithQueueSize, ExecutionChunks};
+
+use crate::{ProcessorError, Result, host::InstanceCreator};
 use chunk_execution_processing::ChunkJournalsProcessingOutput;
-use chunks_splitting::ActorStateHashWithQueueSize;
 use core_processor::common::JournalNote;
 use ethexe_common::{
     BlockHeader, CALL_REPLY_SOFT_LIMIT, OUTGOING_MESSAGES_BYTES_SOFT_LIMIT,
@@ -229,7 +230,7 @@ pub(super) trait RunContext {
     /// The trait method provides a default implementation for a common execution.
     fn handle_chunk_data(
         &self,
-        execution_chunks: &mut ExecutionChunks,
+        execution_chunks: &mut chunks_splitting::ExecutionChunks,
         actor_state: ActorStateHashWithQueueSize,
         queue_type: MessageType,
     ) {
@@ -455,315 +456,6 @@ pub(super) fn states(
             Some(actor_state)
         })
         .collect()
-}
-
-pub(super) mod chunks_splitting {
-    use super::*;
-
-    // An alias introduced for better readability of the chunks splitting steps.
-    type Chunks = Vec<Vec<(ActorId, H256)>>;
-
-    // `prepare_execution_chunks` is not exactly sorting (sorting usually `n*log(n)` this one is `O(n)`),
-    // but rather partitioning into subsets (chunks) of programs with approximately similar queue sizes.
-    pub(super) fn prepare_execution_chunks(
-        ctx: &mut impl RunContext,
-        queue_type: MessageType,
-    ) -> Chunks {
-        let states = ctx.states(queue_type);
-        let mut execution_chunks = ExecutionChunks::new(ctx.inner().chunk_size, states.len());
-
-        for state in states {
-            ctx.handle_chunk_data(&mut execution_chunks, state, queue_type);
-        }
-
-        execution_chunks.arrange_execution_chunks(ctx)
-    }
-
-    /// A helper  struct to bundle actor id, state hash and queue size together
-    /// for easier handling in chunk preparation.
-    pub struct ActorStateHashWithQueueSize {
-        pub actor_id: ActorId,
-        pub hash: H256,
-        pub canonical_queue_size: usize,
-        pub injected_queue_size: usize,
-    }
-
-    impl ActorStateHashWithQueueSize {
-        pub fn new(actor_id: ActorId, state: StateHashWithQueueSize) -> Self {
-            Self {
-                actor_id,
-                hash: state.hash,
-                canonical_queue_size: state.canonical_queue_size as usize,
-                injected_queue_size: state.injected_queue_size as usize,
-            }
-        }
-    }
-
-    /// A helper struct to manage execution chunks during their preparation.
-    pub struct ExecutionChunks {
-        chunk_size: usize,
-        chunks: Chunks,
-    }
-
-    impl ExecutionChunks {
-        fn new(chunk_size: usize, tasks_len: usize) -> Self {
-            let number_of_chunks = tasks_len.div_ceil(chunk_size);
-
-            Self {
-                chunk_size,
-                chunks: vec![vec![]; number_of_chunks],
-            }
-        }
-
-        /// Gets chunk index in chunks tasks queue.
-        pub fn chunk_idx(&self, mq_size: usize) -> usize {
-            // Simplest implementation of chunk partitioning '| 1 | 2 | 3 | 4 | ..'
-            debug_assert_ne!(mq_size, 0);
-            mq_size.min(self.chunks.len()) - 1
-        }
-
-        /// Inserts chunk execution data into the specified chunk index.
-        pub fn insert_into(&mut self, idx: usize, actor_id: ActorId, hash: H256) {
-            if let Some(chunk) = self.chunks.get_mut(idx) {
-                chunk.push((actor_id, hash));
-            } else {
-                panic!(
-                    "Chunk index {idx} out of bounds: chunks number - {}",
-                    self.chunks.len()
-                );
-            }
-        }
-
-        /// Insert chunk execution data into the heaviest chunk (most prior, the last one).
-        pub fn insert_into_heaviest(&mut self, actor_id: ActorId, hash: H256) {
-            if let Some(chunk) = self.chunks.last_mut() {
-                chunk.push((actor_id, hash));
-            } else {
-                panic!("Chunks are empty, cannot insert into heaviest chunk");
-            }
-        }
-
-        /// Arranges execution chunks by merging uneven chunks and reversing their order,
-        /// so the heaviest chunks are processed first.
-        fn arrange_execution_chunks<R: RunContext>(self, ctx: &mut R) -> Chunks {
-            self.chunks
-                .into_iter()
-                // Merge uneven chunks
-                .flatten()
-                // Repartition chunks in reverse order to ensure all chunks have an equal number of elements
-                .rev()
-                .chunks(self.chunk_size)
-                // Convert into vector of vectors
-                .into_iter()
-                .map(|c| {
-                    c.into_iter()
-                        // `check_task_no_run` function isn't used in a common execution, but it's used only for an overlaid one.
-                        // The function is intended to nullify program queues only once before execution. If the queue wasn't nullified
-                        // earlier the function will nullify it and skip spawning the job for the program queue as it's empty. If the queue
-                        // was already nullified, the function will return `false` and the job will be spawned as usual.
-                        // For more info, see impl of the [`OverlaidContext`].
-                        .filter(|&(program_id, _)| !ctx.check_task_no_run(program_id))
-                        .collect()
-                })
-                .collect()
-        }
-    }
-}
-
-mod chunk_execution_spawn {
-    use super::*;
-    use crate::{handling::thread_pool::ThreadPool, host::InstanceWrapper};
-    use ethexe_runtime_common::ProcessQueueContext;
-    use std::sync::LazyLock;
-
-    /// An alias introduced for better readability of the chunks execution steps.
-    pub type ChunkItemOutput = (ActorId, H256, ProgramJournals, u64);
-
-    /// Spawns in the thread pool tasks for each program in the chunk remembering position of the program in the chunk.
-    ///
-    /// Each program receives one (same copy) value of gas allowance, because all programs in the chunk are executed in parallel.
-    /// It means that in the same time unit (!) all programs simultaneously charge gas allowance. If programs were to be
-    /// executed concurrently, then each of the program should have received a reference to the global gas allowance counter
-    /// and charge gas from it concurrently.
-    pub async fn spawn_chunk_execution(
-        ctx: &mut impl RunContext,
-        chunk: Vec<(ActorId, H256)>,
-        queue_type: MessageType,
-    ) -> Result<Vec<ChunkItemOutput>> {
-        struct Executable {
-            queue_type: MessageType,
-            block_info: BlockInfo,
-            promise_policy: PromisePolicy,
-            program_id: ActorId,
-            state_hash: H256,
-            instrumented_code: InstrumentedCode,
-            code_metadata: CodeMetadata,
-            executor: InstanceWrapper,
-            db: Box<dyn CASDatabase>,
-            gas_allowance_for_chunk: u64,
-            promise_out_tx: Option<mpsc::UnboundedSender<Promise>>,
-        }
-
-        fn execute_chunk_item(executable: Executable) -> Result<ChunkItemOutput> {
-            let Executable {
-                queue_type,
-                block_info,
-                promise_policy,
-                program_id,
-                state_hash,
-                instrumented_code,
-                code_metadata,
-                mut executor,
-                db,
-                gas_allowance_for_chunk,
-                promise_out_tx,
-            } = executable;
-
-            let (jn, new_state_hash, gas_spent) = executor.run(
-                db,
-                ProcessQueueContext {
-                    program_id,
-                    state_root: state_hash,
-                    queue_type,
-                    instrumented_code,
-                    code_metadata,
-                    gas_allowance: GasAllowanceCounter::new(gas_allowance_for_chunk),
-                    block_info,
-                    promise_policy,
-                },
-                promise_out_tx,
-            )?;
-            Ok((program_id, new_state_hash, jn, gas_spent))
-        }
-
-        static THREAD_POOL: LazyLock<ThreadPool<Executable, Result<ChunkItemOutput>>> =
-            LazyLock::new(|| ThreadPool::new(execute_chunk_item));
-
-        let gas_allowance_for_chunk = ctx
-            .inner()
-            .gas_allowance_counter
-            .left()
-            .min(CHUNK_PROCESSING_GAS_LIMIT);
-
-        let promise_policy = ctx.promise_policy();
-
-        let block_header = ctx.inner().block_header;
-        let block_info = BlockInfo {
-            height: block_header.height,
-            timestamp: block_header.timestamp,
-        };
-
-        let executables = chunk
-            .into_iter()
-            .map(|(program_id, state_hash)| {
-                let (instrumented_code, code_metadata) = ctx.program_code(program_id)?;
-
-                let executor = ctx.inner().instance_creator.instantiate()?;
-
-                Ok(Executable {
-                    queue_type,
-                    block_info,
-                    promise_policy,
-                    program_id,
-                    state_hash,
-                    instrumented_code,
-                    code_metadata,
-                    executor,
-                    db: ctx.inner().db.cas().clone_boxed(),
-                    gas_allowance_for_chunk,
-                    promise_out_tx: ctx.inner().promise_out_tx.clone(),
-                })
-            })
-            .collect::<Result<Vec<_>>>()?;
-
-        THREAD_POOL.spawn_many(executables).try_collect().await
-    }
-}
-
-mod chunk_execution_processing {
-    use super::*;
-    use crate::handling::run::chunk_execution_spawn::ChunkItemOutput;
-
-    // Aliases introduced for better readability of the chunk journals processing steps.
-    type ProgramChunkJournals = (ActorId, ChunkJournals);
-    type ChunkJournals = Vec<ExtendedJournal>;
-    type ExtendedJournal = (Vec<JournalNote>, MessageType, bool);
-
-    /// Output of the chunk journals processing step.
-    ///
-    /// Chunk journals processing is actually a loop, which can break early.
-    /// The early break must also stop other steps of the caller chunk processing
-    /// function. So to expose the logic in a clear way, the enum is introduced.
-    pub enum ChunkJournalsProcessingOutput {
-        Processed,
-        EarlyBreak,
-    }
-
-    /// Collects journals from all executed programs in the chunk.
-    ///
-    /// The [`chunk_execution_spawn::spawn_chunk_execution`] step spawns tasks for each program in the chunk.
-    /// The loop in the functions handles the output of each task:
-    /// - modifies the state by setting a new state hash calculated by the [`ethexe_runtime_common::RuntimeJournalHandler`]
-    /// - collects journals for later processing
-    /// - tracks the maximum gas spent among all programs in the chunk
-    ///
-    /// Due to the nature of the parallel program queues execution (see [`chunk_execution_spawn::spawn_chunk_execution`] gas allowance clarifications),
-    /// the actual gas allowance spent is actually the maximum among all programs in the chunk, not the sum.
-    pub fn collect_chunk_journals(
-        ctx: &mut impl RunContext,
-        chunk_outputs: Vec<ChunkItemOutput>,
-    ) -> (Vec<ProgramChunkJournals>, u64) {
-        let mut max_gas_spent_in_chunk = 0u64;
-
-        let chunk_journals = chunk_outputs
-            .into_iter()
-            .map(
-                |(program_id, new_state_hash, program_journals, gas_spent)| {
-                    // Handle state updates that occurred during journal processing within the runtime (allocations, pages).
-                    // This should happen before processing the journal notes because `send_dispatch` from another program can modify the state.
-                    ctx.inner_mut().transitions.modify(program_id, |state, _| {
-                        state.hash = new_state_hash;
-                    });
-
-                    max_gas_spent_in_chunk = max_gas_spent_in_chunk.max(gas_spent);
-
-                    (program_id, program_journals)
-                },
-            )
-            .collect();
-
-        (chunk_journals, max_gas_spent_in_chunk)
-    }
-
-    /// Processes journal of each program in the chunk.
-    ///
-    /// The processing is done with [`ethexe_runtime_common::JournalHandler`], which actually sends messages
-    /// generated after executing program queues.
-    ///
-    /// The journals are processed sequentially in the order of programs in the chunk.
-    ///
-    /// The `early_break` closure is intended for overlaid execution mode. The closure is intended to
-    /// nullify queues of receiver programs (if not nullified) until the expected reply is found.
-    /// If it's found, no nullification is done and the processing breaks early.
-    pub fn process_chunk_execution_journals(
-        ctx: &mut impl RunContext,
-        chunk_journals: Vec<ProgramChunkJournals>,
-    ) -> ChunkJournalsProcessingOutput {
-        for (program_id, program_journals) in chunk_journals {
-            for (journal, message_type, call_reply) in program_journals {
-                let break_flag = ctx.break_early(&journal);
-
-                let mut journal_handler = ctx.journal_handler(program_id, message_type, call_reply);
-                core_processor::handle_journal(journal, &mut journal_handler);
-
-                if break_flag {
-                    return ChunkJournalsProcessingOutput::EarlyBreak;
-                }
-            }
-        }
-
-        ChunkJournalsProcessingOutput::Processed
-    }
 }
 
 #[cfg(test)]
